requirements.txt
# Reddit API client
praw>=7.7.0

# HTTP client with rate limiting
requests>=2.31.0
requests-ratelimiter>=0.4.0

# Discord webhook
discord-webhook>=1.3.0

# Environment variables
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0

# CLI
click>=8.1.0

# Caching
requests-cache>=1.1.0
